Recalculate when we are ahead of the reference start point
When traveling besides a gpx track while a recalculation has
occurred the result will sometimes makes detours to get back
to the gpx track. This can result in loops when the route
back to the track actually covers the gpx track itself.

By recalculating when being ahead of the gpx track even
though we are not on it will make sure that those loops
will eventually disappear before we arrive at the gpx track
again.