Turn off screen when idle
When the 'screen turn on' feature is in use the screen turns off
only when it was turned on because of an upcoming turn. When the
user turns on the screen by the power button it will not go off
anymore.

This commit will also turn off the screen when the user is idle
for the same period as set in the 'screen turn on' setting. This
also happens when no navigation is in use (maybe prevent this?).

When this commit is used the 'screen turn on' setting could
use a better name as it became a more generic powersave setting.