Look 'ahead' when searching for new start/end point
When the closest start is found do a check if the location is
already ahead of the current point towards the next point. If
so use the next point as the new start.

This avoids unnecessary routing back to the previous point and
creating loops/u-turns.

The same is done backwards for finding the end point.